v1.9.4 Add As<T> extension for char and related unit tests
Added a new `As<T>` extension method for `char` in the `ParsingHelpers` class, enabling parsing of a `char` into a specified type `T` with a default fallback. This method uses `T.TryParse` with the `char` converted to a string.

Also added a unit test `AsInt_FromChar_ShouldBe` to verify the functionality of the new extension method. Test cases include parsing numeric characters like `'1'` and `'6'` into integers.